move common functions to systemtest/utils.go
also setup collins config in testdata
