+++ management/src/systemtests/cli_test.go
@@ -5,10 +5,8 @@ package systemtests
 import (
 	"fmt"
 	"os"
-	"regexp"
 	"strings"
 	"testing"
-	"time"
 
 	tutils "github.com/contiv/systemtests-utils"
 	"github.com/contiv/vagrantssh"
@@ -39,40 +37,6 @@ var (
 	invalidNodeName  = "invalid-test-node"
 	dummyAnsibleFile = "/tmp/yay"
 )
-
-// XXX: go-check doesn't pass the test's state to the test set/teardown functions.
-// So we have no way to know if a test failed and take some approrpate action.
-// This hack let's me do that for now.
-func (s *CliTestSuite) Assert(c *C, obtained interface{}, checker Checker, args ...interface{}) {
-	if c.Check(obtained, checker, args...) == false {
-		s.failed = true
-		c.FailNow()
-	}
-}
-
-func (s *CliTestSuite) startSerf(c *C, nut vagrantssh.TestbedNode) {
-	out, err := tutils.ServiceStartAndWaitForUp(nut, "serf", 30)
-	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-	c.Logf("serf is running. %s", out)
-}
-
-func (s *CliTestSuite) stopSerf(c *C, nut vagrantssh.TestbedNode) {
-	out, err := tutils.ServiceStop(nut, "serf")
-	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-	c.Logf("serf is stopped. %s", out)
-}
-
-func (s *CliTestSuite) startClusterm(c *C, nut vagrantssh.TestbedNode, timeout int) {
-	out, err := tutils.ServiceStartAndWaitForUp(nut, "clusterm", timeout)
-	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-	c.Logf("clusterm is running. %s", out)
-}
-
-func (s *CliTestSuite) restartClusterm(c *C, nut vagrantssh.TestbedNode) {
-	out, err := tutils.ServiceRestartAndWaitForUp(nut, "clusterm", 30)
-	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-	c.Logf("clusterm is running. %s", out)
-}
 
 func (s *CliTestSuite) SetUpSuite(c *C) {
 	pwd, err := os.Getwd()
@@ -112,13 +76,13 @@ func (s *CliTestSuite) SetUpSuite(c *C) {
 	// come up as it waits on collins container to come up and start serving it's API.
 	// This can take a while, so we wait for cluster-manager
 	// to start with a long timeout here. This way we have this long wait only once.
-	s.startClusterm(c, s.tbn1, 1200)
+	s.restartClusterm(c, s.tbn1, 1200)
 	//provide test ansible playbooks and restart cluster-mgr
 	src := fmt.Sprintf("%s/../demo/files/cli_test/*", pwd)
 	dst := "/etc/default/clusterm/"
 	out, err := s.tbn1.RunCommandWithOutput(fmt.Sprintf("sudo cp -rf %s %s", src, dst))
 	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-	s.restartClusterm(c, s.tbn1)
+	s.restartClusterm(c, s.tbn1, 30)
 }
 
 func (s *CliTestSuite) TearDownSuite(c *C) {
@@ -129,14 +93,6 @@ func (s *CliTestSuite) TearDownSuite(c *C) {
 	s.tbn1 = nil
 	s.tbn2 = nil
 	s.tb.Teardown()
-}
-
-func (s *CliTestSuite) nukeNodeInCollins(c *C, nodeName string) {
-	// Ignore errors here as asset might not exist.
-	out, err := s.tbn1.RunCommandWithOutput(fmt.Sprintf(`curl --basic -u blake:admin:first -d status="Decommissioned" -d reason="test" -X POST http://localhost:9000/api/asset/%s`, nodeName))
-	c.Logf("asset decommission result: %v. Output: %s", err, out)
-	out, err = s.tbn1.RunCommandWithOutput(fmt.Sprintf(`curl --basic -u blake:admin:first -d reason=test -X DELETE http://localhost:9000/api/asset/%s`, nodeName))
-	c.Logf("asset deletion result: %v. Output: %s", err, out)
 }
 
 func (s *CliTestSuite) SetUpTest(c *C) {
@@ -160,7 +116,7 @@ func (s *CliTestSuite) SetUpTest(c *C) {
 		s.nukeNodeInCollins(c, name)
 	}
 
-	s.restartClusterm(c, s.tbn1)
+	s.restartClusterm(c, s.tbn1, 30)
 }
 
 func (s *CliTestSuite) TearDownTest(c *C) {
@@ -206,22 +162,6 @@ func (s *CliTestSuite) TestCommissionDisappearedNode(c *C) {
 	s.assertMatch(c, exptStr, out)
 }
 
-func checkProvisionStatus(tbn1 vagrantssh.TestbedNode, nodeName, exptdStatus string) (string, error) {
-	exptdStr := fmt.Sprintf(`.*"status".*"%s".*`, exptdStatus)
-	return tutils.WaitForDone(func() (string, bool) {
-		cmdStr := fmt.Sprintf("clusterctl node get %s", nodeName)
-		out, err := tbn1.RunCommandWithOutput(cmdStr)
-		if err != nil {
-			return out, false
-			//replace newline with empty string for regex to match properly
-		} else if match, err := regexp.MatchString(exptdStr,
-			strings.Replace(out, "\n", "", -1)); err == nil && match {
-			return out, true
-		}
-		return out, false
-	}, 1*time.Second, 30*time.Second, fmt.Sprintf("node is still not in %q status", exptdStatus))
-}
-
 func (s *CliTestSuite) TestCommissionProvisionFailure(c *C) {
 	// temporarily move the site.yml file to sitmulate a failure
 	pwd, err := os.Getwd()
@@ -243,57 +183,9 @@ func (s *CliTestSuite) TestCommissionProvisionFailure(c *C) {
 	s.Assert(c, err, IsNil, Commentf("output: %s", out))
 }
 
-func (s *CliTestSuite) commissionNode(c *C, nodeName string, nut vagrantssh.TestbedNode) {
-	// provision the node
-	cmdStr := fmt.Sprintf("clusterctl node commission %s", nodeName)
-	out, err := s.tbn1.RunCommandWithOutput(cmdStr)
-	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-	out, err = checkProvisionStatus(s.tbn1, nodeName, "Allocated")
-	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-
-	// verify that site.yml got executed on the node and created the dummy file
-	s.waitForStatToSucceed(c, nut, dummyAnsibleFile)
-}
-
 func (s *CliTestSuite) TestCommissionSuccess(c *C) {
 	nodeName := validNodeNames[0]
 	s.commissionNode(c, nodeName, s.tbn1)
-}
-
-func (s *CliTestSuite) waitForStatToSucceed(c *C, nut vagrantssh.TestbedNode, file string) {
-	out, err := tutils.WaitForDone(func() (string, bool) {
-		cmdStr := fmt.Sprintf("stat -t %s", file)
-		out, err := nut.RunCommandWithOutput(cmdStr)
-		if err != nil {
-			return out, false
-		}
-		return out, true
-	}, 1*time.Second, 10*time.Second, fmt.Sprintf("file %q still doesn't seems to exist", file))
-	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-}
-
-func (s *CliTestSuite) waitForStatToFail(c *C, nut vagrantssh.TestbedNode, file string) {
-	out, err := tutils.WaitForDone(func() (string, bool) {
-		cmdStr := fmt.Sprintf("stat -t %s", file)
-		out, err := nut.RunCommandWithOutput(cmdStr)
-		if err == nil {
-			return out, false
-		}
-		return out, true
-	}, 1*time.Second, 10*time.Second, fmt.Sprintf("file %q still seems to exist", file))
-	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-}
-
-func (s *CliTestSuite) decommissionNode(c *C, nodeName string, nut vagrantssh.TestbedNode) {
-	// decommission the node
-	cmdStr := fmt.Sprintf("clusterctl node decommission %s", nodeName)
-	out, err := s.tbn1.RunCommandWithOutput(cmdStr)
-	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-	out, err = checkProvisionStatus(s.tbn1, nodeName, "Decommissioned")
-	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-
-	// verify that cleanup.yml got executed on the node and deleted the dummy file
-	s.waitForStatToFail(c, nut, dummyAnsibleFile)
 }
 
 func (s *CliTestSuite) TestDecommissionSuccess(c *C) {
@@ -345,22 +237,6 @@ func (s *CliTestSuite) TestDiscoverNodeAlreadyExistError(c *C) {
 	s.assertMatch(c, exptdOut, out)
 }
 
-func (s *CliTestSuite) waitForSerfMembership(c *C, nut vagrantssh.TestbedNode, nodeName, state string) {
-	out, err := tutils.WaitForDone(func() (string, bool) {
-		out, err := nut.RunCommandWithOutput(`serf members`)
-		if err != nil {
-			return out, false
-		}
-		stateStr := fmt.Sprintf(`%s.*%s.*`, nodeName, state)
-		if match, err := regexp.MatchString(stateStr, out); err != nil || !match {
-			return out, false
-		}
-		return out, true
-	}, 1*time.Second, time.Duration(10)*time.Second,
-		fmt.Sprintf("%s's serf membership is not in %s state", nodeName, state))
-	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-}
-
 func (s *CliTestSuite) TestDiscoverSuccess(c *C) {
 	nodeName := validNodeNames[1]
 	nodeAddr := validNodeAddrs[1]
@@ -375,7 +251,7 @@ func (s *CliTestSuite) TestDiscoverSuccess(c *C) {
 	s.waitForSerfMembership(c, s.tbn1, nodeName, "failed")
 
 	// restart clusterm
-	s.restartClusterm(c, s.tbn1)
+	s.restartClusterm(c, s.tbn1, 30)
 
 	// make sure node is not visible in inventory
 	s.getNodeInfoFailureNonExistentNode(c, nodeName)
@@ -412,28 +288,8 @@ func (s *CliTestSuite) TestSetGetGlobalExtraVarsFailureInvalidJSON(c *C) {
 	s.assertMatch(c, exptdOut, out)
 }
 
-func (s *CliTestSuite) getNodeInfoFailureNonExistentNode(c *C, nodeName string) {
-	cmdStr := fmt.Sprintf(`clusterctl node get %s`, nodeName)
-	out, err := s.tbn1.RunCommandWithOutput(cmdStr)
-	s.Assert(c, err, NotNil, Commentf("output: %s", out))
-	exptdOut := fmt.Sprintf(`.*node with name.*%s.*doesn't exists.*`, nodeName)
-	s.assertMatch(c, exptdOut, out)
-}
-
 func (s *CliTestSuite) TestGetNodeInfoFailureNonExistentNode(c *C) {
 	s.getNodeInfoFailureNonExistentNode(c, invalidNodeName)
-}
-
-func (s *CliTestSuite) getNodeInfoSuccess(c *C, nodeName string) {
-	cmdStr := fmt.Sprintf(`clusterctl node get %s`, nodeName)
-	out, err := s.tbn1.RunCommandWithOutput(cmdStr)
-	s.Assert(c, err, IsNil, Commentf("output: %s", out))
-	exptdOut := `.*"monitoring-state":.*`
-	s.assertMultiMatch(c, exptdOut, out, 1)
-	exptdOut = `.*"inventory-state":.*`
-	s.assertMultiMatch(c, exptdOut, out, 1)
-	exptdOut = `.*"configuration-state".*`
-	s.assertMultiMatch(c, exptdOut, out, 1)
 }
 
 func (s *CliTestSuite) TestGetNodeInfoSuccess(c *C) {
@@ -450,17 +306,4 @@ func (s *CliTestSuite) TestGetNodesInfoSuccess(c *C) {
 	s.assertMultiMatch(c, exptdOut, out, 2)
 	exptdOut = `.*"configuration-state".*`
 	s.assertMultiMatch(c, exptdOut, out, 2)
-}
-
-func (s *CliTestSuite) assertMatch(c *C, exptd, rcvd string) {
-	// XXX: the `Matches` checker doesn't match the expression in a multi-line
-	// output so resorting to a regex check here.
-	if match, err := regexp.MatchString(exptd, rcvd); err != nil || !match {
-		s.Assert(c, false, Equals, true, Commentf("output: %s", rcvd))
-	}
-}
-
-func (s *CliTestSuite) assertMultiMatch(c *C, exptd, rcvd string, eMatchCount int) {
-	r := regexp.MustCompile(fmt.Sprintf("(?m)%s", exptd))
-	s.Assert(c, len(r.FindAllString(rcvd, -1)), Equals, eMatchCount, Commentf("output: %s", rcvd))
 }
